Split routes.ts into authorize.ts and token.ts
